Added archivable shared examples
